requirements.txt
akshare>=1.14.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
tenacity>=8.0.0
tqdm>=4.60.0
